gh-658 (refactor) start splitting up janus update class
It was a method with 208 lines before with tremendous complexity. Trying
to make it smaller step by step without breaking functionality.
